# public/swagger/swagger.yaml
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
  Account:
    description: A tenant account (default view)
    example:
      created_at: 2015-09-23T13:51:57-05:00
      created_by: me@goa.design
      href: /accounts/1
      id: 1
      name: test
    properties:
      created_at:
        description: Date of creation
        example: 2015-09-23T13:51:57-05:00
        format: date-time
        type: string
      created_by:
        description: Email of account owner
        example: me@goa.design
        format: email
        type: string
      href:
        description: API href of account
        example: /accounts/1
        type: string
      id:
        description: ID of account
        example: 1
        format: int64
        type: integer
      name:
        description: Name of account
        example: test
        type: string
    required:
    - id
    - href
    - name
    - created_at
    - created_by
    title: 'Mediatype identifier: application/vnd.account+json; view=default'
    type: object
  AccountCollection:
    description: AccountCollection is the media type for an array of Account (default
      view)
    example:
    - created_at: 2015-09-23T13:51:57-05:00
      created_by: me@goa.design
      href: /accounts/1
      id: 1
      name: test
    items:
      $ref: '#/definitions/Account'
    title: 'Mediatype identifier: application/vnd.account+json; type=collection'
    type: array
...
  BottleCollection:
    description: BottleCollection is the media type for an array of Bottle (default
      view)
    example:
    - account:
        created_at: 2015-09-23T13:51:57-05:00
        created_by: me@goa.design
        href: /accounts/1
        id: 1
        name: test
      href: /accounts/1/bottles/1
      id: 1
      links:
        account:
          href: /accounts/1
          id: 1
      name: Number 8
      rating: 3
      varietal: Merlot
      vineyard: Asti
      vintage: 2012
    - account:
        created_at: 2015-09-23T13:51:57-05:00
        created_by: me@goa.design
        href: /accounts/1
        id: 1
        name: test
      href: /accounts/1/bottles/1
      id: 1
      links:
        account:
          href: /accounts/1
          id: 1
      name: Number 8
      rating: 3
      varietal: Merlot
      vineyard: Asti
      vintage: 2012
    - account:
        created_at: 2015-09-23T13:51:57-05:00
        created_by: me@goa.design
        href: /accounts/1
        id: 1
        name: test
      href: /accounts/1/bottles/1
      id: 1
      links:
        account:
          href: /accounts/1
          id: 1
      name: Number 8
      rating: 3
      varietal: Merlot
      vineyard: Asti
      vintage: 2012
    items:
      $ref: '#/definitions/Bottle'
    title: 'Mediatype identifier: application/vnd.bottle+json; type=collection'
    type: array
...
  CreateAccountPayload:
    example:
      name: test
    properties:
      name:
        description: Name of account
        example: test
        type: string
    required:
    - name
    title: CreateAccountPayload
    type: object
...
  RateBottlePayload:
    example:
      rating: 3
    properties:
      rating:
        description: Rating of bottle between 1 and 5
        example: 3
        maximum: 5
        minimum: 1
        type: integer
    required:
    - rating
    title: RateBottlePayload
    type: object
  UpdateAccountPayload:
    example:
      name: test
    properties:
      name:
        description: Name of account
        example: test
        type: string
    required:
    - name
    title: UpdateAccountPayload
    type: object
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1.458609e+09
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1.458609e+09
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error'
    type: object
...
host: localhost:8081
info:
  contact:
    email: admin@goa.design
    name: goa team
    url: http://goa.design
  description: A basic example of a CRUD API implemented with goa
  license:
    name: MIT
    url: https://github.com/goadesign/goa/blob/master/LICENSE
  title: The virtual wine cellar
  version: ""
paths:
...
  /cellar/accounts:
    get:
      description: Retrieve all accounts.
      operationId: account#list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AccountCollection'
      schemes:
      - http
      summary: list account
      tags:
      - account
    post:
      description: Create new account
      operationId: account#create
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/CreateAccountPayload'
      responses:
        "201":
          description: Resource created
          headers:
            Location:
              description: href to created resource
              pattern: /accounts/[0-9]+
              type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: create account
      tags:
      - account
  /cellar/accounts/{accountID}:
    delete:
      operationId: account#delete
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      summary: delete account
      tags:
      - account
    get:
      description: Retrieve account with given id. IDs 1 and 2 pre-exist in the system.
      operationId: account#show
      parameters:
      - description: Account ID
        in: path
        minimum: 1
        name: accountID
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      summary: show account
      tags:
      - account
    put:
      description: Change account name
      operationId: account#update
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: integer
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/UpdateAccountPayload'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      summary: update account
      tags:
      - account
  /cellar/accounts/{accountID}/bottles:
    get:
      description: List all bottles in account optionally filtering by year
      operationId: bottle#list
      parameters:
      - description: Account ID
        in: path
        minimum: 1
        name: accountID
        required: true
        type: integer
      - description: Filter by years
        in: query
        items:
          type: integer
        name: years
        required: false
        type: array
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/BottleCollection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      summary: list bottle
      tags:
      - bottle
    post:
      description: Record new bottle
      operationId: bottle#create
      parameters:
      - description: Account ID
        in: path
        minimum: 1
        name: accountID
        required: true
        type: integer
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/CreateBottlePayload'
      responses:
        "201":
          description: Resource created
          headers:
            Location:
              description: href to created resource
              pattern: ^/accounts/[0-9]+/bottles/[0-9]+$
              type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      summary: create bottle
      tags:
      - bottle
  /cellar/accounts/{accountID}/bottles/{bottleID}:
    delete:
      operationId: bottle#delete
      parameters:
      - description: Account ID
        in: path
        minimum: 1
        name: accountID
        required: true
        type: integer
      - in: path
        name: bottleID
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      summary: delete bottle
      tags:
      - bottle
    get:
      description: Retrieve bottle with given id
      operationId: bottle#show
      parameters:
      - description: Account ID
        in: path
        minimum: 1
        name: accountID
        required: true
        type: integer
      - in: path
        name: bottleID
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Bottle'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      summary: show bottle
      tags:
      - bottle
    patch:
      operationId: bottle#update
      parameters:
      - description: Account ID
        in: path
        minimum: 1
        name: accountID
        required: true
        type: integer
      - in: path
        name: bottleID
        required: true
        type: integer
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/BottlePayload'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      summary: update bottle
      tags:
      - bottle
  /cellar/accounts/{accountID}/bottles/{bottleID}/actions/rate:
    put:
      operationId: bottle#rate
      parameters:
      - description: Account ID
        in: path
        minimum: 1
        name: accountID
        required: true
        type: integer
      - in: path
        name: bottleID
        required: true
        type: integer
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/RateBottlePayload'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      summary: rate bottle
      tags:
      - bottle
  /cellar/accounts/{accountID}/bottles/{bottleID}/watch:
    get:
      description: Retrieve bottle with given id
      operationId: bottle#watch
      parameters:
      - description: Account ID
        in: path
        minimum: 1
        name: accountID
        required: true
        type: integer
      - in: path
        name: bottleID
        required: true
        type: integer
      responses:
        "101":
          description: Switching Protocols
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
      schemes:
      - ws
      summary: watch bottle
      tags:
      - bottle
...
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
...
schemes:
- http
swagger: "2.0"